trezorlib: move mostly-debug methods from device to debuglink
